fix charts date & confirmation delete modal

+++ frontend/src/components/portefeuilles/UploadPortefuille/SavedPtfs.jsx
@@ -10,6 +10,8 @@ import {
   setSelectedPtf,
 } from "../../../redux/slices/BacktestSlice";
 import { notyf } from "../../../utils/notyf";
+import ModalComponent from "../../Modal";
+import DeleteModal from "../../DeleteModal";
 
 const types = ["Actions", "OPCVM"];
 
@@ -19,6 +21,7 @@ const SavedPtfs = ({ selectedPtfs, setSelectedPtfs, show, setShow }) => {
   } = useSelector((state) => state.user);
   const [type, setType] = useState("Actions");
   const [ptf, setPtf] = useState(null);
+  const [isOpen, setIsOpen] = useState(false);
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -49,16 +52,19 @@ const SavedPtfs = ({ selectedPtfs, setSelectedPtfs, show, setShow }) => {
     dispatch(setSelectedPtf(ptf));
   };
 
-  const handleDelete = () => {
-    dispatch(deletePortefeuilles({ ptfs: selectedPtfs }))
-      .unwrap()
-      .then(({ message, portefeuilles }) => {
-        setPtf(null);
-        dispatch(setPortefeuilles(portefeuilles));
-        notyf.success(message);
-      })
-      .catch(() => notyf.error("Error Delete"));
-    setSelectedPtfs([]);
+  const handleDelete = (confirmation) => {
+    if (confirmation) {
+      dispatch(deletePortefeuilles({ ptfs: selectedPtfs }))
+        .unwrap()
+        .then(({ message, portefeuilles }) => {
+          setPtf(null);
+          dispatch(setPortefeuilles(portefeuilles));
+          notyf.success(message);
+          setSelectedPtfs([]);
+        })
+        .catch(() => notyf.error("Error Delete"));
+    }
+    setIsOpen(false);
     console.log("selected ptfs", selectedPtfs);
   };
   useEffect(() => {
@@ -69,41 +75,46 @@ const SavedPtfs = ({ selectedPtfs, setSelectedPtfs, show, setShow }) => {
     .map((ptf) => ptf.name);
 
   return (
-    <Box className="flex flex-wrap gap-3 items-center">
-      <SingleSelect
-        label={"Type"}
-        options={types}
-        value={type}
-        setValue={setType}
-      />
-      <SingleSelect
-        label={"Portefeuilles"}
-        options={portefeuilles}
-        value={ptf}
-        setValue={setPtf}
-      />
-      <Button
-        variant="contained"
-        className="min-w-[115px] flex gap-2 items-center"
-        color="primary"
-        size="small"
-        id="valider-btn"
-        onClick={handleValider}
-        disabled={!ptf}
-      >
-        Valider <CheckSquare size={18} />
-      </Button>
-      <Button
-        variant="contained"
-        color="error"
-        size="small"
-        className="min-w-[115px] flex gap-2 items-center"
-        onClick={handleDelete}
-        disabled={false}
-      >
-        Supprimer <Trash size={18} />
-      </Button>
-    </Box>
+    <>
+      <Box className="flex flex-wrap gap-3 items-center">
+        <SingleSelect
+          label={"Type"}
+          options={types}
+          value={type}
+          setValue={setType}
+        />
+        <SingleSelect
+          label={"Portefeuilles"}
+          options={portefeuilles}
+          value={ptf}
+          setValue={setPtf}
+        />
+        <Button
+          variant="contained"
+          className="min-w-[115px] flex gap-2 items-center"
+          color="primary"
+          size="small"
+          id="valider-btn"
+          onClick={handleValider}
+          disabled={!ptf}
+        >
+          Valider <CheckSquare size={18} />
+        </Button>
+        <Button
+          variant="contained"
+          color="error"
+          size="small"
+          className="min-w-[115px] flex gap-2 items-center"
+          onClick={() => setIsOpen(true)}
+          disabled={!ptf}
+        >
+          Supprimer <Trash size={18} />
+        </Button>
+      </Box>
+      <ModalComponent open={isOpen} handleClose={() => setIsOpen(false)}>
+        <DeleteModal handleDeleteConfirmation={handleDelete} />
+      </ModalComponent>
+    </>
   );
 };
